prepare handle_data in jsonl backend
