+++ src/piddiplatsch/handles/jsonl_backend.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from typing import Any
 
 from piddiplatsch.config import config
-from piddiplatsch.handles.base import HandleBackend
+from piddiplatsch.handles.base import HandleBackend, prepare_handle_data
 
 
 class JsonlHandleBackend(HandleBackend):
@@ -19,8 +19,9 @@ class JsonlHandleBackend(HandleBackend):
 
     def add(self, pid: str, record: dict[str, Any]) -> None:
         """Add or overwrite a PID record (naive append)."""
+        handle_data = prepare_handle_data(record)
         with self.path.open("a") as f:
-            f.write(json.dumps({"pid": pid, "record": record}) + "\n")
+            f.write(json.dumps({"pid": pid, "record": handle_data}) + "\n")
 
     def get(self, pid: str) -> dict[str, Any] | None:
         """Retrieve a PID record. Return None if not found."""
